agent/requirements.txt
# Core dependencies for the Telegram Agent
google-adk>=1.18.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
psycopg2-binary>=2.9.9

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
